[CLI] Add "ti changelog" command to print changelog of current version (#1153)
* better make_changelog.py

* [CLI] add "ti changelog" command to print changelog of current version

* [skip ci] enforce code format

* [skip ci] Apply suggestions from code review


* [skip ci] enforce code format

* [skip ci] Update python/build.py
