+++ python/MANIFEST.in
@@ -2,6 +2,7 @@ include MANIFEST.in
 include *.txt
 include *.py
 include *.cfg
+include taichi/*.md
 include taichi/assets/*
 include taichi/lib/*.so
 include taichi/lib/*.pyd
